Move FileAnnotationTab to own package, fix localization
