+++ src/main/java/org/jabref/gui/entryeditor/fileannotationtab/FileAnnotationTabController.java
@@ -1,4 +1,4 @@
-package org.jabref.gui.entryeditor;
+package org.jabref.gui.entryeditor.fileannotationtab;
 
 import javax.inject.Inject;
 
@@ -23,12 +23,17 @@ import org.fxmisc.easybind.EasyBind;
 
 public class FileAnnotationTabController extends AbstractController<FileAnnotationTabViewModel> {
 
-    @FXML ComboBox<String> files;
-    @FXML ListView<FileAnnotationViewModel> annotationList;
-    @FXML Label author;
+    @FXML
+    ComboBox<String> files;
+    @FXML
+    ListView<FileAnnotationViewModel> annotationList;
+    @FXML
+    Label author;
 
-    @Inject private FileAnnotationCache fileAnnotationCache;
-    @Inject private BibEntry entry;
+    @Inject
+    private FileAnnotationCache fileAnnotationCache;
+    @Inject
+    private BibEntry entry;
 
     @FXML
     public void initialize() {
@@ -54,9 +59,9 @@ public class FileAnnotationTabController extends AbstractController<FileAnnotati
                     HBox details = new HBox();
                     details.getStyleClass().setAll("details");
                     Text page = new Text();
-                    page.setText(Localization.lang("Page: ") + Integer.toString(annotation.getPage()));
+                    page.setText(Localization.lang("Page") + ": " + Integer.toString(annotation.getPage()));
                     Text author = new Text();
-                    author.setText(Localization.lang("Author: ") + annotation.getAuthor());
+                    author.setText(Localization.lang("Author") + ": " + annotation.getAuthor());
                     details.getChildren().addAll(page, author);
 
                     node.getChildren().addAll(text, details);
