+++ src/main/java/org/jabref/gui/entryeditor/fileannotationtab/FileAnnotationViewModel.java
@@ -1,4 +1,4 @@
-package org.jabref.gui.entryeditor;
+package org.jabref.gui.entryeditor.fileannotationtab;
 
 import java.util.Optional;
